replaces confirmation toast in article.show

+++ vue/src/views/Article/Show.vue
@@ -3,20 +3,17 @@ import { ref } from 'vue'
 
 import { storeToRefs } from 'pinia'
 import { useRoute } from 'vue-router'
-import { useToast } from 'vue-toastification'
 
 import { useUserStore } from '@/stores/user'
 
 import CommentService from '@/services/CommentService'
 import ArticleService from '@/services/ArticleService'
+import ToastService from '@/services/ToastService'
 
-import ConfirmationToast from '@/components/atoms/ConfirmationToast.vue'
 import AttachmentCard from '@/components/AttachmentCard.vue'
 import CommentForm from '@/components/atoms/CommentForm.vue'
 import ItemLine from '@/components/atoms/ItemLine.vue'
 import ContentRenderer from './ContentRenderer.vue'
-
-const toast = useToast()
 
 const userStore = useUserStore()
 const { hasPermission } = storeToRefs(userStore)
@@ -33,24 +30,10 @@ const loadFromServer = () => {
 }
 
 const deleteComment = comment => {
-  toast.clear()
-  const content = {
-    component: ConfirmationToast,
-    props: {
-      message: 'Kommentar wirklich entfernen?',
-    },
-    listeners: {
-      granted: () => {
-        CommentService.deleteComment(comment).then(data => {
-          loadFromServer()
-        })
-      },
-    },
-  }
-  toast(content, {
-    timeout: false,
-    icon: false,
-    closeButton: false,
+  ToastService.confirm('Kommentar wirklich entfernen?', () => {
+    CommentService.deleteComment(comment).then(data => {
+      loadFromServer()
+    })
   })
 }
